veth: Write setupVeth in a separate subpackage
This makes the logical structure match the one from  other "backends"
such as sriov.

Meanwhile, also bring setInterfaceUp into pkg/veth/veth.go as it is a
helper function that will only be used by the veth backend.
